rename pubsub as redis-pubsub
