+++ index.js
@@ -4,18 +4,18 @@ const http = require('http');
 const defaultPort = 80;
 const port = process.env.MS_PORT || defaultPort;
 const app = express();
-const pubsub = require("./src/pubsub").postHandler;
+const pubsub = require("./src/redis-pubsub").postHandler;
 const bodyParser = require("body-parser");
 const jsonParser = bodyParser.json();
 const server = http.createServer(app);
-const displayConnections = require("./src/messages/display-connections.js");
-const datastore = require("./src/db/redis/datastore.js");
-const watch = require("./src/messages/watch.js");
-const presence = require("./src/presence/presence.js");
+const displayConnections = require("./src/messages/display-connections");
+const datastore = require("./src/db/redis/datastore");
+const watch = require("./src/messages/watch");
+const presence = require("./src/presence/presence");
 const pkg = require("./package.json");
 const podname = process.env.podname;
-const logger = require("./src/logger.js");
-const gcs = require("./src/version-compare/gcs.js");
+const logger = require("./src/logger");
+const gcs = require("./src/version-compare/gcs");
 const querystring = require("querystring");
 const url = require("url");
